# services/payment/api/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Service API
  description: |
    API for managing payment processing and payment lifecycle.

    This service handles:
    - Payment initiation and authorization
    - Payment status tracking
    - Payment provider integration (Stripe-like)
    - Webhook handling for payment provider callbacks
    - Refund processing
  version: 1.0.0
  contact:
    name: GoCommerce Team
    email: team@gocommerce.com
servers:
  - url: http://localhost:8084
    description: Local development server
  - url: http://payment-service:8084
    description: Docker environment
...
paths:
...
  /payments:
    post:
      summary: Create payment
      description: Initiates a new payment for an order (idempotent)
      operationId: createPayment
      tags:
        - payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '200':
          description: Payment already exists (idempotent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List payments
      description: Returns a list of payments with optional filtering
      operationId: listPayments
      tags:
        - payments
      parameters:
        - name: order_id
          in: query
          required: false
          description: Filter by order ID
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          description: Filter by payment status
          schema:
            type: string
            enum: [pending, authorized, captured, failed, cancelled, refunded]
        - name: limit
          in: query
          required: false
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Payments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/{paymentId}:
    get:
      summary: Get payment
      description: Retrieves a specific payment by ID
      operationId: getPayment
      tags:
        - payments
      parameters:
        - name: paymentId
          in: path
          required: true
          description: Payment ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/{paymentId}/capture:
    post:
      summary: Capture payment
      description: Captures an authorized payment (charge the customer)
      operationId: capturePayment
      tags:
        - payments
      parameters:
        - name: paymentId
          in: path
          required: true
          description: Payment ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment captured successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Invalid request or payment cannot be captured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/{paymentId}/cancel:
    post:
      summary: Cancel payment
      description: Cancels a pending or authorized payment
      operationId: cancelPayment
      tags:
        - payments
      parameters:
        - name: paymentId
          in: path
          required: true
          description: Payment ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Invalid request or payment cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/{paymentId}/refund:
    post:
      summary: Refund payment
      description: Refunds a captured payment (full refund)
      operationId: refundPayment
      tags:
        - payments
      parameters:
        - name: paymentId
          in: path
          required: true
          description: Payment ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '200':
          description: Payment refunded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Invalid request or payment cannot be refunded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /webhooks/payment-provider:
    post:
      summary: Payment provider webhook
      description: Receives webhook callbacks from payment provider (e.g., Stripe)
      operationId: handleWebhook
      tags:
        - webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEvent'
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '400':
          description: Invalid webhook payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    CreatePaymentRequest:
      type: object
      required:
        - order_id
        - amount
        - currency
      properties:
        order_id:
          type: string
          format: uuid
          description: Order ID this payment is for
        amount:
          type: number
          format: double
          description: Payment amount
          minimum: 0.01
          example: 99.99
        currency:
          type: string
          description: Currency code (ISO 4217)
          minLength: 3
          maxLength: 3
          example: "USD"
        idempotency_key:
          type: string
          description: Idempotency key to prevent duplicate payments
          example: "order-123-payment-attempt-1"
        payment_method:
          type: string
          description: Payment method type
          enum: [card, bank_transfer, wallet]
          default: card
        metadata:
          type: object
          description: Additional metadata
          additionalProperties: true
    PaymentResponse:
      type: object
      required:
        - id
        - order_id
        - amount
        - currency
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Payment ID
        order_id:
          type: string
          format: uuid
          description: Order ID
        amount:
          type: number
          format: double
          description: Payment amount
        currency:
          type: string
          description: Currency code
        status:
          type: string
          enum: [pending, authorized, captured, failed, cancelled, refunded]
          description: Payment status
        payment_method:
          type: string
          description: Payment method type
        provider_payment_id:
          type: string
          description: Payment ID from payment provider
        failure_reason:
          type: string
          description: Reason for failure (if failed)
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    PaymentListResponse:
      type: object
      required:
        - payments
        - total
        - limit
        - offset
      properties:
        payments:
          type: array
          items:
            $ref: '#/components/schemas/PaymentResponse'
        total:
          type: integer
          description: Total number of payments matching filters
        limit:
          type: integer
          description: Limit used in query
        offset:
          type: integer
          description: Offset used in query
    RefundRequest:
      type: object
      properties:
        reason:
          type: string
          description: Reason for refund
          example: "Customer requested refund"
    WebhookEvent:
      type: object
      required:
        - event_type
        - payment_id
      properties:
        event_type:
          type: string
          enum: [payment.authorized, payment.captured, payment.failed, payment.refunded]
          description: Type of webhook event
        payment_id:
          type: string
          description: Payment ID from provider
        data:
          type: object
          description: Event data from provider
          additionalProperties: true
    WebhookResponse:
      type: object
      required:
        - received
      properties:
        received:
          type: boolean
          description: Whether webhook was received successfully
          example: true
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "PAYMENT_NOT_FOUND"
        message:
          type: string
          description: Error message
          example: "Payment not found"
        details:
          type: object
          description: Additional error details
          additionalProperties: true