# services/payment/internal/adapters/http/generated.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package http

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for CreatePaymentRequestPaymentMethod.
const (
	BankTransfer CreatePaymentRequestPaymentMethod = "bank_transfer"
	Card         CreatePaymentRequestPaymentMethod = "card"
	Wallet       CreatePaymentRequestPaymentMethod = "wallet"
)

// Valid indicates whether the value is a known member of the CreatePaymentRequestPaymentMethod enum.
func (e CreatePaymentRequestPaymentMethod) Valid() bool {
	switch e {
	case BankTransfer:
		return true
	case Card:
		return true
	case Wallet:
		return true
	default:
		return false
	}
}

// Defines values for PaymentResponseStatus.
const (
	PaymentResponseStatusAuthorized PaymentResponseStatus = "authorized"
	PaymentResponseStatusCancelled  PaymentResponseStatus = "cancelled"
	PaymentResponseStatusCaptured   PaymentResponseStatus = "captured"
	PaymentResponseStatusFailed     PaymentResponseStatus = "failed"
	PaymentResponseStatusPending    PaymentResponseStatus = "pending"
	PaymentResponseStatusRefunded   PaymentResponseStatus = "refunded"
)

// Valid indicates whether the value is a known member of the PaymentResponseStatus enum.
func (e PaymentResponseStatus) Valid() bool {
	switch e {
	case PaymentResponseStatusAuthorized:
		return true
	case PaymentResponseStatusCancelled:
		return true
	case PaymentResponseStatusCaptured:
		return true
	case PaymentResponseStatusFailed:
		return true
	case PaymentResponseStatusPending:
		return true
	case PaymentResponseStatusRefunded:
		return true
	default:
		return false
	}
}

// Defines values for WebhookEventEventType.
const (
	PaymentAuthorized WebhookEventEventType = "payment.authorized"
	PaymentCaptured   WebhookEventEventType = "payment.captured"
	PaymentFailed     WebhookEventEventType = "payment.failed"
	PaymentRefunded   WebhookEventEventType = "payment.refunded"
)

// Valid indicates whether the value is a known member of the WebhookEventEventType enum.
func (e WebhookEventEventType) Valid() bool {
	switch e {
	case PaymentAuthorized:
		return true
	case PaymentCaptured:
		return true
	case PaymentFailed:
		return true
	case PaymentRefunded:
		return true
	default:
		return false
	}
}

// Defines values for ListPaymentsParamsStatus.
const (
	ListPaymentsParamsStatusAuthorized ListPaymentsParamsStatus = "authorized"
	ListPaymentsParamsStatusCancelled  ListPaymentsParamsStatus = "cancelled"
	ListPaymentsParamsStatusCaptured   ListPaymentsParamsStatus = "captured"
	ListPaymentsParamsStatusFailed     ListPaymentsParamsStatus = "failed"
	ListPaymentsParamsStatusPending    ListPaymentsParamsStatus = "pending"
	ListPaymentsParamsStatusRefunded   ListPaymentsParamsStatus = "refunded"
)

// Valid indicates whether the value is a known member of the ListPaymentsParamsStatus enum.
func (e ListPaymentsParamsStatus) Valid() bool {
	switch e {
	case ListPaymentsParamsStatusAuthorized:
		return true
	case ListPaymentsParamsStatusCancelled:
		return true
	case ListPaymentsParamsStatusCaptured:
		return true
	case ListPaymentsParamsStatusFailed:
		return true
	case ListPaymentsParamsStatusPending:
		return true
	case ListPaymentsParamsStatusRefunded:
		return true
	default:
		return false
	}
}

// CreatePaymentRequest defines model for CreatePaymentRequest.
type CreatePaymentRequest struct {
	// Amount Payment amount
	//
	// Example: 99.99
	Amount float64 `json:"amount"`

	// Currency Currency code (ISO 4217)
	//
	// Example: USD
	Currency string `json:"currency"`

	// IdempotencyKey Idempotency key to prevent duplicate payments
	//
	// Example: order-123-payment-attempt-1
	IdempotencyKey *string `json:"idempotency_key,omitempty"`

	// Metadata Additional metadata
	Metadata *map[string]interface{} `json:"metadata,omitempty"`

	// OrderId Order ID this payment is for
	OrderId openapi_types.UUID `json:"order_id"`

	// PaymentMethod Payment method type
	PaymentMethod *CreatePaymentRequestPaymentMethod `json:"payment_method,omitempty"`
}

// CreatePaymentRequestPaymentMethod Payment method type
type CreatePaymentRequestPaymentMethod string

// Error defines model for Error.
type Error struct {
	// Details Additional error details
	Details *map[string]interface{} `json:"details,omitempty"`

	// Error Error code
	//
	// Example: PAYMENT_NOT_FOUND
	Error string `json:"error"`

	// Message Error message
	//
	// Example: Payment not found
	Message string `json:"message"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	// Service Example: payment-service
	Service string `json:"service"`

	// Status Example: healthy
	Status    string     `json:"status"`
	Timestamp *time.Time `json:"timestamp,omitempty"`
}

// PaymentListResponse defines model for PaymentListResponse.
type PaymentListResponse struct {
	// Limit Limit used in query
	Limit int `json:"limit"`

	// Offset Offset used in query
	Offset   int               `json:"offset"`
	Payments []PaymentResponse `json:"payments"`

	// Total Total number of payments matching filters
	Total int `json:"total"`
}

// PaymentResponse defines model for PaymentResponse.
type PaymentResponse struct {
	// Amount Payment amount
	Amount float64 `json:"amount"`

	// CreatedAt Creation timestamp
	CreatedAt time.Time `json:"created_at"`

	// Currency Currency code
	Currency string `json:"currency"`

	// FailureReason Reason for failure (if failed)
	FailureReason *string `json:"failure_reason,omitempty"`

	// Id Payment ID
	Id openapi_types.UUID `json:"id"`

	// OrderId Order ID
	OrderId openapi_types.UUID `json:"order_id"`

	// PaymentMethod Payment method type
	PaymentMethod *string `json:"payment_method,omitempty"`

	// ProviderPaymentId Payment ID from payment provider
	ProviderPaymentId *string `json:"provider_payment_id,omitempty"`

	// Status Payment status
	Status PaymentResponseStatus `json:"status"`

	// UpdatedAt Last update timestamp
	UpdatedAt time.Time `json:"updated_at"`
}

// PaymentResponseStatus Payment status
type PaymentResponseStatus string

// RefundRequest defines model for RefundRequest.
type RefundRequest struct {
	// Reason Reason for refund
	//
	// Example: Customer requested refund
	Reason *string `json:"reason,omitempty"`
}

// WebhookEvent defines model for WebhookEvent.
type WebhookEvent struct {
	// Data Event data from provider
	Data *map[string]interface{} `json:"data,omitempty"`

	// EventType Type of webhook event
	EventType WebhookEventEventType `json:"event_type"`

	// PaymentId Payment ID from provider
	PaymentId string `json:"payment_id"`
}

// WebhookEventEventType Type of webhook event
type WebhookEventEventType string

// WebhookResponse defines model for WebhookResponse.
type WebhookResponse struct {
	// Received Whether webhook was received successfully
	//
	// Example: true
	Received bool `json:"received"`
}

// ListPaymentsParams defines parameters for ListPayments.
type ListPaymentsParams struct {
	// OrderId Filter by order ID
	OrderId *openapi_types.UUID `form:"order_id,omitempty" json:"order_id,omitempty"`

	// Status Filter by payment status
	Status *ListPaymentsParamsStatus `form:"status,omitempty" json:"status,omitempty"`

	// Limit Maximum number of results
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`

	// Offset Number of results to skip
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// ListPaymentsParamsStatus defines parameters for ListPayments.
type ListPaymentsParamsStatus string

// CreatePaymentJSONRequestBody defines body for CreatePayment for application/json ContentType.
type CreatePaymentJSONRequestBody = CreatePaymentRequest

// RefundPaymentJSONRequestBody defines body for RefundPayment for application/json ContentType.
type RefundPaymentJSONRequestBody = RefundRequest

// HandleWebhookJSONRequestBody defines body for HandleWebhook for application/json ContentType.
type HandleWebhookJSONRequestBody = WebhookEvent

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// HealthCheck Health check
	// (GET /health)
	HealthCheck(w http.ResponseWriter, r *http.Request)
	// ListPayments List payments
	// (GET /payments)
	ListPayments(w http.ResponseWriter, r *http.Request, params ListPaymentsParams)
	// CreatePayment Create payment
	// (POST /payments)
	CreatePayment(w http.ResponseWriter, r *http.Request)
	// GetPayment Get payment
	// (GET /payments/{paymentId})
	GetPayment(w http.ResponseWriter, r *http.Request, paymentId openapi_types.UUID)
	// CancelPayment Cancel payment
	// (POST /payments/{paymentId}/cancel)
	CancelPayment(w http.ResponseWriter, r *http.Request, paymentId openapi_types.UUID)
	// CapturePayment Capture payment
	// (POST /payments/{paymentId}/capture)
	CapturePayment(w http.ResponseWriter, r *http.Request, paymentId openapi_types.UUID)
	// RefundPayment Refund payment
	// (POST /payments/{paymentId}/refund)
	RefundPayment(w http.ResponseWriter, r *http.Request, paymentId openapi_types.UUID)
	// HandleWebhook Payment provider webhook
	// (POST /webhooks/payment-provider)
	HandleWebhook(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// HealthCheck Health check
// (GET /health)
func (_ Unimplemented) HealthCheck(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ListPayments List payments
// (GET /payments)
func (_ Unimplemented) ListPayments(w http.ResponseWriter, r *http.Request, params ListPaymentsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// CreatePayment Create payment
// (POST /payments)
func (_ Unimplemented) CreatePayment(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// GetPayment Get payment
// (GET /payments/{paymentId})
func (_ Unimplemented) GetPayment(w http.ResponseWriter, r *http.Request, paymentId openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// CancelPayment Cancel payment
// (POST /payments/{paymentId}/cancel)
func (_ Unimplemented) CancelPayment(w http.ResponseWriter, r *http.Request, paymentId openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// CapturePayment Capture payment
// (POST /payments/{paymentId}/capture)
func (_ Unimplemented) CapturePayment(w http.ResponseWriter, r *http.Request, paymentId openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// RefundPayment Refund payment
// (POST /payments/{paymentId}/refund)
func (_ Unimplemented) RefundPayment(w http.ResponseWriter, r *http.Request, paymentId openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// HandleWebhook Payment provider webhook
// (POST /webhooks/payment-provider)
func (_ Unimplemented) HandleWebhook(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// HealthCheck operation middleware
func (siw *ServerInterfaceWrapper) HealthCheck(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.HealthCheck(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListPayments operation middleware
func (siw *ServerInterfaceWrapper) ListPayments(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params ListPaymentsParams

	// ------------- Optional query parameter "order_id" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "order_id", r.URL.Query(), &params.OrderId, runtime.BindQueryParameterOptions{Type: "string", Format: "uuid"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "order_id"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "order_id", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "status" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "status", r.URL.Query(), &params.Status, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "status"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "status", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "offset", r.URL.Query(), &params.Offset, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "offset"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "offset", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListPayments(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreatePayment operation middleware
func (siw *ServerInterfaceWrapper) CreatePayment(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreatePayment(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetPayment operation middleware
func (siw *ServerInterfaceWrapper) GetPayment(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "paymentId" -------------
	var paymentId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "paymentId", chi.URLParam(r, "paymentId"), &paymentId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "paymentId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetPayment(w, r, paymentId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CancelPayment operation middleware
func (siw *ServerInterfaceWrapper) CancelPayment(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "paymentId" -------------
	var paymentId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "paymentId", chi.URLParam(r, "paymentId"), &paymentId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "paymentId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CancelPayment(w, r, paymentId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CapturePayment operation middleware
func (siw *ServerInterfaceWrapper) CapturePayment(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "paymentId" -------------
	var paymentId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "paymentId", chi.URLParam(r, "paymentId"), &paymentId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "paymentId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CapturePayment(w, r, paymentId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RefundPayment operation middleware
func (siw *ServerInterfaceWrapper) RefundPayment(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "paymentId" -------------
	var paymentId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "paymentId", chi.URLParam(r, "paymentId"), &paymentId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "paymentId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RefundPayment(w, r, paymentId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// HandleWebhook operation middleware
func (siw *ServerInterfaceWrapper) HandleWebhook(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.HandleWebhook(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/health", wrapper.HealthCheck)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/payments", wrapper.ListPayments)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/payments", wrapper.CreatePayment)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/payments/{paymentId}", wrapper.GetPayment)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/payments/{paymentId}/capture", wrapper.CapturePayment)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/payments/{paymentId}/cancel", wrapper.CancelPayment)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/payments/{paymentId}/refund", wrapper.RefundPayment)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/webhooks/payment-provider", wrapper.HandleWebhook)
	})

	return r
}